feat: upgrade SubtitleFlow to v0.5.0 evidence-driven release system (#2)
* feat: add centralized editorial policy model

* feat: model evidence grades and conflicts

* feat: classify subtitle semantic roles

* feat: add deterministic bilingual block geometry

* feat: add bilingual reconciliation contract

* feat: extend workfile and review provenance models

* feat: enforce editorial policy in semantic review

* fix: preserve legacy projects without editorial config

* feat: emit semantic risk signals from alignment

* feat: classify ASS events by semantic role

* feat: reconcile bilingual source evidence during prepare

* feat: compile reconciled deterministic bilingual blocks

* feat: define mode-aware Kiwi Collector geometry

* feat: sync packaged Kiwi Collector geometry

* feat: add synthetic libass renderer QA

* feat: separate structural static and renderer QA

* feat: generate first-class release audit artifacts

* feat: attach auditable release evidence artifacts

* test: add evidence-driven release regressions

* fix: stop treating source style names as protection authority

* ci: add full SubtitleFlow verification workflow

* docs: align OpenCode alignment contract with reconciliation

* docs: teach JP skill editorial policy and evidence rules

* docs: align clean skill with semantic role and policy contracts

* docs: align release QA skill with renderer and audit pipeline

* docs: give semantic editor one Editorial Context contract

* docs: align orchestrator with editorial and release contracts

* docs: align QA reviewer with evidence and renderer risks

* ci: retain lint diagnostics as artifacts

* test: add real-world bilingual renderer specimen

* test: document real-world evidence specimen

* docs: document evidence-driven subtitle release system

* ci: run functional gates before baseline lint audit

* ci: allow one-shot verified patch application

* chore: stage verified release-system fixes

* fix: resolve release-system regression findings

* chore: stage v0.5 docs and config sync

* fix: resolve release-system regression findings

* chore: clean up CI patch scaffolding, format repo, and harden renderer QA

---------

+++ src/subtitleflow/timecode.py
@@ -23,7 +23,7 @@ def parse_ass_time(value: str) -> int:
 def format_ass_time(milliseconds: int) -> str:
     if milliseconds < 0:
         raise ValidationError("Negative ASS timestamp")
-    total_cs = int(round(milliseconds / 10))
+    total_cs = round(milliseconds / 10)
     hours, rem = divmod(total_cs, 360_000)
     minutes, rem = divmod(rem, 6_000)
     seconds, centiseconds = divmod(rem, 100)
